apps/intermediary/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
arq>=0.26.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
google-cloud-storage>=2.10.0